TST: move Series.map tests to tests/series/methods/test_map.py
